+++ projects/plugins/jetpack/class-jetpack-xmlrpc-methods.php
@@ -44,6 +44,8 @@ class Jetpack_XMLRPC_Methods {
 	/**
 	 * Returns what features are available. Uses the slug of the module files.
 	 *
+	 * @deprecated 13.9
+	 * @see Jetpack_Core_Json_Api_Endpoints::get_features_available()
 	 * @return array
 	 */
 	public static function features_available() {
@@ -59,6 +61,8 @@ class Jetpack_XMLRPC_Methods {
 	/**
 	 * Returns what features are enabled. Uses the slug of the modules files.
 	 *
+	 * @deprecated 13.9
+	 * @see Jetpack_Core_Json_Api_Endpoints::get_features_enabled()
 	 * @return array
 	 */
 	public static function features_enabled() {